address comments, fix full screen bug for pagination
